extractor_temp/career_navigator/requirements.txt
spacy==3.7.2
nltk==3.8.1
PyGithub==2.1.1
pandas==2.1.4
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
python-dotenv==1.0.0
transformers==4.36.2
torch==2.1.2
huggingface-hub==0.20.2
pdfplumber==0.10.3
tabula-py==2.9.0
camelot-py[cv]==0.11.0
